Separate webpack configurations to common, development and production
